requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-docx>=1.1.0
anthropic>=0.20.0
openai>=1.10.0
Pillow>=10.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
